Dumbing down the delete, moving the component into the note component.
